LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/anchor-download-intercept-reject_no-currententrychange-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: boo

Harness Error (FAIL), message = Unhandled rejection: boo

FAIL event and promise ordering for <a download> intercepted by passing a rejected promise to intercept() assert_equals: error objects must match: error object for transition.finished rejected did not match the one for navigateerror expected object "Error: boo" but got object "UnknownError: Uncaught Error: boo"
